move import_metadata_from_file to datasource_state.py
